handle exception for grocery

+++ src/services/grocery.ts
@@ -2,15 +2,25 @@ import { UUID } from "crypto";
 import { IGrocery, IGroceryQuery } from "../interfaces/grocery";
 import { GroceryModel } from "../models/grocery";
 import loggerWithNameSpace from "../utils/logger";
+import { NotFoundError } from "../error/NotFoundError";
+import { BadRequestError } from "../error/BadRequestError";
 
 const logger = loggerWithNameSpace("service: grocery");
 
 export function createGrocery(data: IGrocery) {
+  logger.info("Creating grocery");
+
   return GroceryModel.create(data);
 }
 
 export async function getGroceriesForAdmin(filter: IGroceryQuery) {
   logger.info("Getting all groceries for admin");
+
+  const data = await GroceryModel.get(filter);
+
+  if (filter.id && !data[0]) {
+    throw new NotFoundError("Grocery does not exist");
+  }
 
   return await GroceryModel.get(filter);
 }
@@ -19,6 +29,16 @@ export async function getGroceries(filter: IGroceryQuery) {
   logger.info("Getting all groceries");
 
   let data = await GroceryModel.get(filter);
+
+  if (filter.id && !data[0]) {
+    throw new NotFoundError("Grocery does not exist");
+  }
+
+  if (filter.id) {
+    return data[0];
+  }
+
+  // filter out deleted and out-of-stock groceries
   data = data.filter((grocery) => {
     return grocery.quantity > 0 && !grocery.deletedAt;
   });
@@ -30,20 +50,45 @@ export async function getGroceries(filter: IGroceryQuery) {
       description: grocery.description,
       price: grocery.price,
       quantity: grocery.quantity,
+      deletedAt: grocery.deletedAt,
     };
   });
 }
 
 export async function updateGrocery(id: UUID, data: IGrocery) {
+  logger.info("Updating grocery");
+
+  const existingGrocery = (await GroceryModel.get({ id }))[0];
+
+  if (!existingGrocery || existingGrocery.deletedAt) {
+    throw new NotFoundError("Grocery does not exist");
+  }
+
   return await GroceryModel.update(id, data);
 }
 
 export async function updateQuantity(id: UUID, quantity: number) {
   logger.info("Updating grocery quantity");
 
+  if (quantity < 0) {
+    throw new BadRequestError("Quantity cannot be negative");
+  }
+
+  const existingGrocery = (await GroceryModel.get({ id }))[0];
+
+  if (!existingGrocery || existingGrocery.deletedAt) {
+    throw new NotFoundError("Grocery does not exist");
+  }
+
   return await GroceryModel.updateQuantity(id, quantity);
 }
 
 export async function deleteGrocery(id: UUID) {
+  const existingGrocery = (await GroceryModel.get({ id }))[0];
+
+  if (!existingGrocery || existingGrocery.deletedAt) {
+    throw new NotFoundError("Grocery does not exist");
+  }
+
   return await GroceryModel.delete(id);
 }
